+++ src/style/ImpactStories.less
@@ -25,7 +25,7 @@
 
 .cardSeperator {
     width: 100%;
-    height: 100%;
+    height: fit-content;
 
     padding: 2.5%;
     color: black;
